make _legacy version match what is in 5.0
