+++ libraries/chain/CMakeLists.txt
@@ -94,6 +94,7 @@ add_library( eosio_chain
              transaction.cpp
              block.cpp
              block_header.cpp
+             block_header_state.cpp
              block_header_state_legacy.cpp
              block_state_legacy.cpp
              fork_database.cpp
